Moves LFW dataset class to datadefs directory. Serves as example for further datasets migrations
