+++ cyphercat/datadefs/__init__.py
@@ -1,1 +1,3 @@
+
+from .lfw_dataset import *
 from .libri_dataset import *
